refactor(model): Add Game interface and abstract BoardGame
