+++ src/main/java/edu/ntnu/idi/idatt/filehandler/BoardFileHandlerGson.java
@@ -7,11 +7,11 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.google.gson.JsonPrimitive;
 import edu.ntnu.idi.idatt.model.Board;
-import edu.ntnu.idi.idatt.model.LadderAction;
-import edu.ntnu.idi.idatt.model.PortalAction;
-import edu.ntnu.idi.idatt.model.SlideAction;
-import edu.ntnu.idi.idatt.model.Tile;
-import edu.ntnu.idi.idatt.model.interfaces.TileAction;
+import edu.ntnu.idi.idatt.model.tile.LadderAction;
+import edu.ntnu.idi.idatt.model.tile.PortalAction;
+import edu.ntnu.idi.idatt.model.tile.SlideAction;
+import edu.ntnu.idi.idatt.model.tile.Tile;
+import edu.ntnu.idi.idatt.model.tile.TileAction;
 import java.io.File;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
