+++ src/main/java/edu/ntnu/idi/idatt/model/tile/PortalAction.java
@@ -1,28 +1,31 @@
-package edu.ntnu.idi.idatt.model;
+package edu.ntnu.idi.idatt.model.tile;
 
-import edu.ntnu.idi.idatt.model.interfaces.TileAction;
+import edu.ntnu.idi.idatt.model.Board;
+import edu.ntnu.idi.idatt.model.Player;
 import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionPerformValidator;
 import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionSetDescriptionValidator;
 import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionSetDestinationTileIdValidator;
 import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionSetIdentifierValidator;
+
 /**
- * <h3>SlideAction class</h3>
+ * <h3>PortalAction class</h3>
  *
- * <p>This class represents a slide action, which is a tile action that moves the player to a
- * different tile. It contains a destination tile ID and a description.
+ * <p>This class represents a Portal action, which is a tile action that moves the player to a
+ * random tile. It contains a destination tile ID and a description.
  */
-public class SlideAction implements TileAction {
+public class PortalAction implements TileAction {
   private String identifier;
   private int destinationTileId;
   private String description;
+
   /**
-   * Constructor for SlideAction class.
+   * Constructor for PortalAction class.
    *
    * @param identifier The identifier of the action.
    * @param destinationTileId The ID of the destination tile.
    * @param description The description of the action.
    */
-  public SlideAction(String identifier, int destinationTileId, String description) {
+  public PortalAction(String identifier, int destinationTileId, String description) {
     setIdentifier(identifier);
     setDestinationTileId(destinationTileId);
     setDescription(description);
@@ -90,7 +93,7 @@ public class SlideAction implements TileAction {
   @Override
   public void setDescription(String description) {
     tileActionSetDescriptionValidator(description);
-    
+
     this.description = description;
   }
 
@@ -102,8 +105,8 @@ public class SlideAction implements TileAction {
    */
   @Override
   public void perform(Player player, Board board) {
-      tileActionPerformValidator(player, board);
+    tileActionPerformValidator(player, board);
 
-      player.placeOnTile(board.getTile(this.destinationTileId));
+    player.placeOnTile(board.getTile(this.destinationTileId));
   }
 }
