+++ src/main/java/edu/ntnu/idi/idatt/model/tile/TileAction.java
@@ -1,4 +1,4 @@
-package edu.ntnu.idi.idatt.model.interfaces;
+package edu.ntnu.idi.idatt.model.tile;
 
 import edu.ntnu.idi.idatt.model.Board;
 import edu.ntnu.idi.idatt.model.Player;
