+++ src/main/java/edu/ntnu/idi/idatt/model/tile/WormholeAction.java
@@ -1,30 +1,31 @@
-package edu.ntnu.idi.idatt.model;
+package edu.ntnu.idi.idatt.model.tile;
 
-import edu.ntnu.idi.idatt.model.interfaces.TileAction;
+import edu.ntnu.idi.idatt.model.Board;
+import edu.ntnu.idi.idatt.model.Player;
 import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionPerformValidator;
 import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionSetDescriptionValidator;
 import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionSetDestinationTileIdValidator;
 import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionSetIdentifierValidator;
 
 /**
- * <h3>PortalAction class</h3>
+ * <h3>WormholeAction class</h3>
  *
- * <p>This class represents a Portal action, which is a tile action that moves the player to a
- * random tile. It contains a destination tile ID and a description.
+ * <p>This class represents a Wormhole action, which is a tile action that moves the player to a
+ * random  wormhole tile. It contains a destination tile ID and a description.
  */
-public class PortalAction implements TileAction {
+public class WormholeAction implements TileAction {
   private String identifier;
   private int destinationTileId;
   private String description;
 
   /**
-   * Constructor for PortalAction class.
+   * Constructor for WormholeAction class.
    *
    * @param identifier The identifier of the action.
    * @param destinationTileId The ID of the destination tile.
    * @param description The description of the action.
    */
-  public PortalAction(String identifier, int destinationTileId, String description) {
+  public WormholeAction(String identifier, int destinationTileId, String description) {
     setIdentifier(identifier);
     setDestinationTileId(destinationTileId);
     setDescription(description);
