+++ src/main/java/edu/ntnu/idi/idatt/observer/BoardGameObserver.java
@@ -1,7 +1,7 @@
 package edu.ntnu.idi.idatt.observer;
 
 import edu.ntnu.idi.idatt.model.Player;
-import edu.ntnu.idi.idatt.model.interfaces.TileAction;
+import edu.ntnu.idi.idatt.model.tile.TileAction;
 
 /**
  * An interface for observers that wish to be notified about board game events.
