+++ src/test/java/edu/ntnu/idi/idatt/model/BoardGameTest.java
@@ -1,6 +1,7 @@
 package edu.ntnu.idi.idatt.model;
 
 import edu.ntnu.idi.idatt.factory.BoardFactory;
+import edu.ntnu.idi.idatt.model.tile.Tile;
 import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
