+++ src/test/java/edu/ntnu/idi/idatt/model/DiceTest.java
@@ -1,5 +1,7 @@
 package edu.ntnu.idi.idatt.model;
 
+import edu.ntnu.idi.idatt.model.dice.Dice;
+import edu.ntnu.idi.idatt.model.dice.Die;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
